+++ src/t8_geometry/t8_geometry_base.hxx
@@ -69,7 +69,7 @@ public:
   virtual void        t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const = 0;
 
   /**
@@ -84,7 +84,7 @@ public:
   virtual void        t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const = 0;
 
   /** Update a possible internal data buffer for per tree data.
