+++ src/t8_geometry/t8_geometry_implementations/t8_geometry_occ.hxx
@@ -115,7 +115,7 @@ public:
   t8_geom_evaluate (t8_cmesh_t cmesh,
                     t8_gloidx_t gtreeid,
                     const double *ref_coords,
-                    const int num_coords,
+                    const size_t num_coords,
                     double out_coords[3]) const;
 
   /**
@@ -131,7 +131,7 @@ public:
   t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                             t8_gloidx_t gtreeid,
                             const double *ref_coords,
-                            const int num_coords,
+                            const size_t num_coords,
                             double *jacobian) const;
